SendImage Patch
The sending of images works correctly --> This function may plausibly be used for other functionality however it has only been tested on smaller image files. Handling of files is awkwardly implemented, it will require a typename id e.g. "IMAGE" or "TEXT" etc. similar to that used in HTTP Requests.